# [Perf] Simplify the R1CS > Assignment conversion