The next (after the independent https://github.com/ProvableHQ/snarkVM/pull/2620) chapter in the synthesis optimization saga.

The conversion from `R1CS` to the `Assignment` needed for `ConstraintSynthesizer::generate_constraints` is very costly due to the allocations involved. However, most of it is actually unnecessary, and despite my worries that it had been done in order to reduce memory use, skipping it actually leads to lower utilization of RAM.

The performance impact, measured with the same synthesis as in the aforementioned other PR, is as follows:
- runtime **-21%**
- total allocs **and** deallocs **-18%** each
- peak RAM **-43%**

As for the benchmark from the aforementioned other PR:
```
CheckDeployment for do  time:   [847.87 ms 848.42 ms 848.94 ms]
                        change: [-35.625% -34.795% -34.243%] (p = 0.00 < 0.05)
                        Performance has improved.
```
However, this one also has an impact on another new benchmark from that PR:
```
CheckDeployment for transfer_private
                        time:   [139.91 ms 140.01 ms 140.20 ms]
                        change: [-10.901% -9.3525% -7.7967%] (p = 0.00 < 0.05)
                        Performance has improved.
```
(there is also a new `transfer_public` benchmark, but it's not as stable on my machine, returning results ranging from -3% to +3%)